backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
dj-database-url
gunicorn
django-cors-headers==3.7.0
pandas
scikit-learn
numpy
python-dotenv
openai